requirements.txt
asgiref==3.3.1
cffi==1.14.5
cryptography==3.4.6
Django==3.1.7
django-cors-headers==3.7.0
django-environ==0.4.5
djangorestframework==3.12.2
djangorestframework_simplejwt
environ==1.0
jwt==1.2.0
mysqlclient==2.0.3
pycparser==2.20
PyJWT==2.0.1
pytz==2021.1
sqlparse==0.4.1
